+++ resource/application_hook/default_settings_hook.py
@@ -1,11 +1,18 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2015 ftrack
 
-import ftrack
+import ftrack_api
 
 
 class DefaultExportSettings(object):
     '''Return default values for export settings.'''
+
+    def __init__(self, session, *args, **kwargs):
+        super(DefaultExportSettings, self).__init__(
+            *args, **kwargs
+        )
+
+        self.session = session
 
     def get(self, event):
         '''Return default settings.
@@ -22,21 +29,24 @@ class DefaultExportSettings(object):
         }
 
     def register(self):
+
         '''Register hook.'''
-        ftrack.EVENT_HUB.subscribe(
+        self.session.event_hub.subscribe(
             'topic=ftrack.connect.nuke-studio.get-default-settings',
             self.get
         )
 
 
-def register(registry, **kw):
-    '''Register hooks for default settings.'''
-
-    # Validate that registry is instance of ftrack.Registry, if not
-    # return early since the register method probably is called
-    # from the new API.
-    if not isinstance(registry, ftrack.Registry):
+def register(session, **kw):
+    '''Register plugin. Called when used as an plugin.'''
+    # Validate that session is an instance of ftrack_api.Session. If not,
+    # assume that register is being called from an old or incompatible API and
+    # return without doing anything.
+    if not isinstance(session, ftrack_api.session.Session):
         return
 
-    plugin = DefaultExportSettings()
+    plugin = DefaultExportSettings(
+        session
+    )
+
     plugin.register()