+++ source/ftrack_connect_nuke_studio/ui/crew.py
@@ -210,10 +210,10 @@ class NukeCrew(QtWidgets.QDialog):
                 if not entity:
                     continue
 
-                if entity._type == 'task':
-                    context['shot'].append(entity.getId())
-                elif entity._type == 'component':
-                    component_ids.append(entity.getId())
+                if entity.entity_type == 'Shot':
+                    context['shot'].append(entity.get('id'))
+                elif entity.entity_type == 'Component':
+                    component_ids.append(entity.get('id'))
 
         if component_ids:
             components = session.query(