+++ doc/release/release_notes.rst
@@ -8,6 +8,15 @@ Release Notes
 *************
 
 .. release:: Upcoming
+    .. change:: fixed
+        :tags: Export project
+
+        Show an error dialog if the img asset type does not exist in the server.
+
+    .. change:: new
+        :tags: API
+
+        Remove dependencies on the ftrack legacy API where possible
 
     .. change:: new
         :tags: Template, Structure