+++ doc/release/release_notes.rst
@@ -8,6 +8,12 @@ Release Notes
 *************
 
 .. release:: Upcoming
+
+    .. change:: fixed
+        :tags: API
+
+        Nuke 11 not supported.
+
     .. change:: fixed
         :tags: Export project