+++ doc/release/release_notes.rst
@@ -8,9 +8,10 @@ Release Notes
 *************
 
 .. release:: Upcoming
-    .. change:: fixed
-        Fail gracefully if not able to find rv install location in linux.
 
+    .. change:: fixed
+
+        Plugin outputs error if installation location is not found for RV under Linux.
 
     .. change:: fixed